[Feat] #82 - 공고 상세 페이지 조회 API 연결

+++ Terning-iOS/Terning-iOS/Network/TargetType/AnnouncementsTargetType.swift
@@ -9,7 +9,7 @@ import Foundation
 import Moya
 
 enum AnnouncementsTargetType {
-    case getAnnouncements(internshipAnnouncementId: Double)
+    case getAnnouncements(internshipAnnouncementId: Int)
 }
 
 extension AnnouncementsTargetType: TargetType {
